Hello world on Lambda via C++
Okay, so not the best hello world. If the JSON payload is more than 42
bytes in length an error is raised; otherwise, the payload is returned
back unmodified.

All the build commands are manual still:

* Have to build the aws-lambda-cpp project separately and include that
  path in the install prefix of the next command.
* mkdir build && cd build && cmake .. -DCMAKE_EXPORT_COMPILE_COMMANDS=1 -DCMAKE_BUILD_TYPE=Debug
  -DCMAKE_INSTALL_PREFIX=~/.local/lambda-install
* make && make aws-lambda-package-demo
* terraform apply -auto-approve

Making slash commands has to be done via an API POST, so I also included
a small poetry setup that is capable of that. Eventually I'll replace
this with C++ code.